int float

+++ Project_wsl_2026/Software/main_tennsy4.0/lib/IR/compute/compute.hpp
@@ -5,10 +5,12 @@
 #include "IR.hpp"
 
 // 回り込む角度を出力
-int get_IR_mawarikomi_deg();
+int16_t get_IR_mawarikomi_deg();
+
 // IRボールに近寄る角度を出力
-int get_IR_follow_deg(int follow_target_offset);
+int16_t get_IR_follow_deg(int16_t follow_target_offset);
+
 // IRボールの角度に比例する角度を出力
-int get_IR_hirei_deg(double a);
+int16_t get_IR_hirei_deg(double a);
 
 #endif
