+++ Project_wsl_2026/Software/main_tennsy4.0/lib/IR/serial/serial.hpp
@@ -5,15 +5,19 @@
 #include "IR.hpp"
 
 // シリアル定義・開始
-void IR_init(HardwareSerial *serial, int baudrate);
+void IR_init(HardwareSerial *serial, uint32_t baudrate);
+
 // 更新
 void IR_update();
 
 // ボールが存在するかどうか
 bool is_IR_exist();
+
 // ボールの方向を出力
-int get_IR_deg();
+int16_t get_IR_deg();
+
 // ボールまでの距離を出力
-int get_IR_distance();
+int16_t get_IR_distance();
 
 #endif
+
