+++ Project_wsl_2026/Software/main_tennsy4.0/lib/LINE/compute/compute.hpp
@@ -14,13 +14,14 @@ bool get_LINE_side_left();  // 左サイドの状況を取得
 bool get_LINE_side_back();  // 後サイドの状況を取得
 
 // === 角度取得 ===
-// ノーマル版の角度を取得
-int get_LINE_deg();
-// メモリ版の角度を取得
-int get_LINE_memory_deg();
+// ノーマル版の角度を取得 (-1.0f で未使用)
+int16_t get_LINE_deg();
+// メモリ版の角度を取得 (-1.0f で未使用)
+int16_t get_LINE_memory_deg();
 
 // === 最初に反応したピン（共通） ===
-int get_first_triggered_pin(); // 最初に反応したピンを取得
+// -1 で未使用
+int8_t get_first_triggered_pin(); 
 
 // === ハーフアウト検知（共通） ===
 bool is_LINE_half_out(); // ハーフアウトしたかどうか
