+++ Project_wsl_2026/Software/main_tennsy4.0/lib/Motors/compute/motors-boost_ver.hpp
@@ -4,15 +4,14 @@
 #include <Arduino.h>
 #include "Motors.hpp"
 
-
 // それぞれのチャンネルに接続したモータの角度を指定
-void motors_init(int deg_1ch, int deg_2ch, int deg_3ch, int deg_4ch);
+void motors_init(int16_t deg_1ch, int16_t deg_2ch, int16_t deg_3ch, int16_t deg_4ch);
 
 // 指定した方向に動かす（PD制御付き）
-void motors_move(int deg, int abs_power);
+void motors_move(int16_t deg, int16_t abs_power);
 
 // PD制御のみをする
-void motors_only_PD(int max_pd_power);
+void motors_only_PD(int16_t max_pd_power);
 
 // すべてのモータを停止させる
 void motors_break();
@@ -21,9 +20,9 @@ void motors_break();
 
 // degはロボット全体として進みたい方向
 // powerはその方向へ進むときの合力の大きさ
-void compute_motor_power(int deg, int power);
+void compute_motor_power(int16_t deg, int16_t power);
 
 // index番目のモータの合力の大きさを返す
-double get_motor_power(int index);
+float get_motor_power(int8_t index);
 
 #endif
